pyproject.toml
[tool.poetry]
name = "chronos"
version = "1.0.0"
description = "Manage DND Party timezones"
authors = ["PurpleMyst <PurpleMyst@users.noreply.github.com>"]
license = "GPL-3.0-only"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.4.1"
structlog = "^20.1.0"
colorama = "^0.4.3"
HumanTime = "^0.1.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 79